cdvae_property_models/pyproject.toml
[project]
name = "cdvae-property-models"
version = "1.0.0"
description = "CDVAE property model evaluation for crystal structures"
authors = [{ name = "Nikita Kazeev", email = "kazeevn@gmail.com" }]
license = {text = "Apache License (2.0)"}
requires-python = ">=3.12, <3.13"
dependencies = [
  "numpy >=1.26.0, <2",
  "torch ==2.10.0",
  "torch-scatter ==2.1.2",
  "torch-sparse ==0.6.18",
  "torch-geometric >=2.7.0",
  "pytorch-lightning >=2.4.0",
  "lightning >=2.4.0",
  "hydra-core >=1.3.2",
  "omegaconf >=2.3.0",
  "pymatgen ==2025.10.7",
  "pyxtal >=1.1.3",
  "scikit-learn >=1.5.0",
  "p-tqdm >=1.4.2",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 5 - Production/Stable",
    "Environment :: GPU :: NVIDIA CUDA",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Physics"
]
[project.optional-dependencies]
publish = ["build", "twine"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cdvae_property_models"]

[tool.hatch.build]
include = [
  "src/cdvae_property_models/prop_models/**/*",
  "src/cdvae_property_models/pl_modules/gemnet/gemnet-dT.json",
]

[tool.uv.extra-build-dependencies]
torch-scatter = ["torch"]
torch-sparse = ["torch"]
